requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
distlib
dj-database-url
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-rest-auth
djangorestframework
drf-extra-fields
EasyProcess
entrypoint2
filelock
filetype
idna
Jinja2
libmagic
Markdown
MarkupSafe
MouseInfo
mss
numpy
oauthlib
opencv-contrib-python
Pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
PyJWT
PyMsgBox
pyperclip
pyproject
PyRect
pyscreenshot
PyScreeze
python-decouple
python-magic
python3-openid
pytweening
pytz
requests
requests-oauthlib
six
sqlparse
typing_extensions
tzdata
urllib3
virtualenv